fix: rename dialog components dir and improve typing of components
