--- src/lib/components/dialogs/actions.svelte.ts
@@ -1,20 +1,0 @@
-import type { Action } from 'svelte/action'
-
-export const portal: Action<HTMLElement, string> = (el, target = 'body') => {
-	$effect(() => {
-		const targetEl = document.querySelector(target)
-
-		if (!targetEl) {
-			throw new Error(`Target element "${target}" not found`)
-		}
-
-		targetEl.appendChild(el)
-		el.hidden = false
-
-		return () => {
-			if (el.parentNode) {
-				el.parentNode.removeChild(el)
-			}
-		}
-	})
-}
